# packages/contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fake News Detection API",
    "description": "Local-first API adapter for fake-news and media analysis.",
    "version": "0.3.0"
  },
  "paths": {
...
    "/v1/analyses/text": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Analyze Text",
        "operationId": "analyze_text_v1_analyses_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyses/url": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Analyze Url",
        "operationId": "analyze_url_v1_analyses_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeUrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Get Analysis",
        "operationId": "get_analysis_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AnalysisResponse"
                    },
                    {
                      "$ref": "#/components/schemas/PendingAnalysisResponse"
                    }
                  ],
                  "title": "Response Get Analysis V1 Analyses  Analysis Id  Get"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        }
      }
    },
...
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Events",
        "operationId": "get_job_events_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "after_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After Sequence"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Events stream of job progress.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisResponse": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "input_type": {
            "type": "string",
            "enum": [
              "text",
              "url",
              "file",
              "unknown"
            ],
            "title": "Input Type"
          },
          "media_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "image",
                  "audio",
                  "video"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Type"
          },
          "extracted_text": {
            "type": "string",
            "title": "Extracted Text"
          },
          "cleaned_text": {
            "type": "string",
            "title": "Cleaned Text"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "media_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Metadata"
          },
          "extraction_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "style_signal": {
            "type": "string",
            "enum": [
              "LOW_STYLE_RISK",
              "HIGH_STYLE_RISK",
              "UNKNOWN",
              "ERROR"
            ],
            "title": "Style Signal"
          },
          "style_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Confidence"
          },
          "style_scope_reliable": {
            "type": "boolean",
            "title": "Style Scope Reliable"
          },
          "style_word_count": {
            "type": "integer",
            "title": "Style Word Count"
          },
          "style_minimum_word_count": {
            "type": "integer",
            "title": "Style Minimum Word Count"
          },
          "style_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Warning"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "final_verdict": {
            "type": "string",
            "title": "Final Verdict"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Confidence"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analysis_id",
          "status",
          "input_type",
          "extracted_text",
          "cleaned_text",
          "source_url",
          "style_signal",
          "style_confidence",
          "style_scope_reliable",
          "style_word_count",
          "style_minimum_word_count",
          "verification",
          "final_verdict",
          "confidence",
          "reason",
          "created_at",
          "completed_at",
          "request_id",
          "trace_id"
        ],
        "title": "AnalysisResponse"
      },
      "AnalyzeTextRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Text"
          },
          "deep_check": {
            "type": "boolean",
            "title": "Deep Check",
            "default": false
          },
          "max_length": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 8192.0,
                "minimum": 128.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Length"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "AnalyzeTextRequest"
      },
      "AnalyzeUrlRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "deep_check": {
            "type": "boolean",
            "title": "Deep Check",
            "default": false
          },
          "max_length": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 8192.0,
                "minimum": 128.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Length"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url"
        ],
        "title": "AnalyzeUrlRequest"
      },
      "ApiErrorResponse": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error_code",
          "message",
          "request_id",
          "trace_id"
        ],
        "title": "ApiErrorResponse"
      },
...
      "ExtractionMetadata": {
        "properties": {
          "media_type": {
            "type": "string",
            "enum": [
              "image",
              "audio",
              "video"
            ],
            "title": "Media Type"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "channels": {
            "additionalProperties": true,
            "type": "object",
            "title": "Channels"
          },
          "timings_ms": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Timings Ms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "media_type"
        ],
        "title": "ExtractionMetadata"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "JobError"
      },
      "JobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "validating",
              "preprocessing",
              "extracting",
              "cleaning",
              "style_analysis",
              "searching",
              "fetching_evidence",
              "verifying",
              "deciding",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "current_stage": {
            "type": "string",
            "enum": [
              "queued",
              "validating",
              "preprocessing",
              "extracting",
              "cleaning",
              "style_analysis",
              "searching",
              "fetching_evidence",
              "verifying",
              "deciding",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Current Stage"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "analysis_id",
          "status",
          "progress",
          "current_stage",
          "message",
          "created_at",
          "updated_at",
          "request_id",
          "trace_id"
        ],
        "title": "JobResponse"
      },
...
      "PendingAnalysisResponse": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "input_type": {
            "type": "string",
            "const": "file",
            "title": "Input Type"
          },
          "media_type": {
            "type": "string",
            "enum": [
              "image",
              "audio",
              "video"
            ],
            "title": "Media Type"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analysis_id",
          "status",
          "input_type",
          "media_type",
          "job_id",
          "created_at",
          "request_id",
          "trace_id"
        ],
        "title": "PendingAnalysisResponse"
      },
...
      "VerificationResponse": {
        "properties": {
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "evidence_quality": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "HIGH",
                  "MEDIUM",
                  "LOW"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Quality"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          },
          "evidence_summary": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Summary"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceSummaryItem"
            },
            "type": "array",
            "title": "Evidence"
          },
          "web_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Web Context"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "verdict",
          "evidence_quality",
          "explanation",
          "recommendation"
        ],
        "title": "VerificationResponse"
      }
    }
  }
}